requirements.txt
git+git://github.com/dbca-wa/dbca-utils@1.0.0
git+https://github.com/dbca-wa/ledger#egg=ledger
psycopg2-binary==2.7.7
Django<1.11.0
django-extensions==1.9.8
git+https://github.com/dbca-wa/dpaw-utils.git@0.3a16#egg=dpaw-utils
git+https://github.com/django-oscar/django-oscar.git@17a1f6b17aba19e3c0b095e90acc654d463e0cf9#egg=django-oscar
social-auth-app-django==2.1.0
git+https://github.com/scottp-dpaw/social-core.git@email_fix#egg=social-auth-core
coverage==4.3.1
coveralls==1.1
reportlab==3.5.26
django_bootstrap3==7.1.0
django-braces>=1.8.1
django-datatables-view==1.13.0
django-reversion==3.0.0
django-reversion-compare==0.8.6
git+https://github.com/bruth/django-preserialize.git@11f9e822250ea3374d5cfba477a00ca886689d9f#egg=django-preserialize
django-countries==3.4.1
django-cron==0.5.0
django-dynamic-fixture==1.9.1
openpyxl==2.4.8
datapackage==0.8.1
jsontableschema==0.10.1
python-dateutil==2.6.1
py4j==0.10.2.1
djangorestframework==3.5.0
djangorestframework-gis==0.11
djangorestframework-csv==2.0.0
pycountry==17.1.8
six>=1.10.0
django-ical==1.4
django-taggit==0.21.3
geojson==1.3.3
unicodecsv==0.14.1
PyPDF2==1.26.0
django-extra-fields==0.9
django-crispy-forms==1.7.0
django-reset-migrations==0.3.1
django-ckeditor==5.4.0
djangorestframework-datatables==0.4.0
django-confy==1.0.4
xlwt==1.3.0
xlsxwriter==1.1.2
gunicorn==19.9.0
dj-static==0.0.6
Pillow==4.2.1
mixer==6.1.3
django-dirtyfields==1.3.1
